feat: 컴파일 절대경로 시작점 src폴더로 설정
