+++ src/LandingPage/index.js
@@ -1,0 +1,14 @@
+import React from "react";
+import {MainSection, FirstSection, Sections} from "./architecture"
+
+const LandingPage = function () {
+	return (
+		<>
+		<MainSection/>
+		<FirstSection/>
+		<Sections/>
+		</>
+	);
+}
+
+export default LandingPage
