+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client";
 import GlobalStyle from "./style-initialization/global-style";
 import ResetStyle from "./style-initialization/reset-style";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import LandingPage from "./pages/landing-page";
+import LandingPage from "LandingPage";
 import LoginPage from "./pages/login-page";
 import SignupPage from "./pages/signup-page";
 
@@ -18,7 +18,7 @@ root.render(
         <Route path="/" element={<LandingPage />}></Route>
         <Route path="/signup" element={<SignupPage />}></Route>
         <Route path="/login" element={<LoginPage />}></Route>
-		<Route path="/folder" element={<LandingPage />}></Route>
+        <Route path="/folder" element={<LandingPage />}></Route>
       </Routes>
     </BrowserRouter>
   </React.StrictMode>
